refactor(chat): extract session lifecycle (~1k lines) into two part files
The session-lifecycle methods were a large, contiguous, field-free, non-faked
cluster — moved verbatim (zero behaviour change) into two part-file extensions,
split to respect the ~500-line cap:

- chat_service_session_load.dart — _loadLastSession, getSessionsForId,
  getSessions, loadSession (load + listing).
- chat_service_session_manage.dart — renameSession, updateSessionDescription,
  forkFromMessage, startNewChat (management + new chat).

chat_service.dart 8,912 → 7,898 lines. Across this thinning pass it's gone
10,506 → 7,898 (−2,608, ~25%), with cohesive concerns now in 11 part files.

flutter analyze clean; full unit suite (1348) + parity gate green.
