+++ lib/services/chat_service.dart
@@ -93,6 +93,8 @@ part 'chat/chat_service_group_realism_helpers.dart';
 part 'chat/chat_service_history.dart';
 part 'chat/chat_service_group_membership.dart';
 part 'chat/chat_service_reprocess.dart';
+part 'chat/chat_service_session_load.dart';
+part 'chat/chat_service_session_manage.dart';
 
 // Internal flag to signal a cancellation request for realism evaluation.
 // This is a file-scope flag to avoid needing to thread state through the
@@ -3676,1022 +3678,6 @@ class ChatService extends ChangeNotifier {
     });
   }
 
-  Future<void> _loadLastSession() async {
-    if (_activeCharacter == null && _activeGroup == null) return;
-
-    // Get sessions from DB
-    List<Session> sessions;
-    if (_activeGroup != null) {
-      sessions = await _db.getSessionsForGroup(_activeGroup!.id);
-    } else if (_activeCharacter?.dbId != null) {
-      sessions = await _db.getSessionsForCharacter(_activeCharacter!.dbId!);
-    } else {
-      return;
-    }
-
-    if (sessions.isEmpty) {
-      debugPrint('[ChatService] _loadLastSession: No previous sessions found');
-      // Ensure no stale fork/parent state remains from a prior character/group.
-      _parentSessionId = null;
-      _forkIndex = null;
-      // Expression runtime (manual/caches) reset on no-prior-session to prevent bleed (new for step4, matches fork/parent hygiene).
-      _expressionService.resetForFreshChat();
-      // Time (secondary config: passage, weekday anchors, turns, day/time scalars) reset for fresh 0-session/new-group paths.
-      // Prevents bleed of advanced time from prior 1:1 into fresh groups (cross-check vs needs bugfix reset hygiene).
-      // Nsfw (cooldown/arousal) reset for same (incomplete zeroing of nsfw on 0-session/new-group was a prior hygiene issue).
-      // Lorebook triggers/depth reset for same (incomplete zeroing of lore on 0-session/new-group was a prior hygiene pattern to avoid).
-      // See "keep reset blocks in sync" (setActiveGroup, startNewChat 1:1+group (now explicit in both), load* , setActive* all must hit this; now includes needs/chaos/... + leaves (see CLAUDE.md for full; incomplete zeroing now complete) + " ; now complete in all group/0-session/new-chat hygiene)" ; incomplete zeroing now complete).
-      // (cross-ref setActiveCharacter:1572)
-      _timeService.resetForFreshChat();
-      _nsfwService.resetForFreshChat();
-      _lorebookScanner.resetLorebookTriggerState();
-      _activeObjectives = [];
-      _messagesSinceLastCheck = 0;
-      _isCheckingCompletion =
-          false; // zero in _loadLast empty early return (0-session path hygiene)
-      _summaryPaused =
-          false; // explicit secondary zero for _summaryPaused (symmetric; _loadLast empty early return 0-session)
-      _isSummaryGenerating =
-          false; // secondary zero in _loadLast empty (0-session for summary flag)
-      _userMessagesSinceLastPeriodicEval = 0;
-      _isExtractingFacts =
-          false; // secondary fact flag + counter zero in _loadLast empty early return (0-session path hygiene; fact_extraction)
-      _isEvolvingCharacter = false;
-      _evolutionStatus = '';
-      _evolutionError =
-          ''; // explicit evo flag/status/error zero in _loadLast empty early return (0-session path hygiene; evolution_service (stateless or prompt-only; no reset calls needed))
-      return;
-    }
-
-    // Sessions are already sorted descending by createdAt
-    final lastSession = sessions.first;
-    _currentSessionId = lastSession.id;
-    _authorNote = lastSession.authorNote;
-    _authorNoteStrength = lastSession.authorNoteDepth;
-    _summary = lastSession.summary ?? '';
-    _summaryLastIndex = lastSession.summaryLastIndex ?? 0;
-    _sessionName = lastSession.name;
-    _sessionDescription = lastSession.description;
-    _parentSessionId = lastSession.parentSession;
-    _forkIndex = lastSession.forkIndex;
-    // Relationship scalars + migration/tier calc now via service (keeps load parity).
-    // Migration: scale old scores (±150) to new range (±300). (Card-seed bypass: fresh V2.5
-    // ext seeds use seedFromCardV2OrExt plain clamp in setActiveCharacter/startNew 1:1 paths;
-    // this migrate path is *only* for legacy persisted sessions. See card-seed notes at the two
-    // ext-seed sites + relationship_service + full keep-sync lists + setActiveCharacter:1572.)
-    _relationshipService.loadScalars(
-      affectionScore: _relationshipService.migrateShortTermScore(
-        lastSession.affectionScore,
-      ),
-      longTermScore: _relationshipService.migrateLongTermScore(
-        lastSession.longTermScore,
-      ),
-      trustLevel: lastSession.trustLevel,
-      activeFixation: lastSession.activeFixation,
-      fixationLifespan: lastSession.fixationLifespan,
-      spatialStance: lastSession.spatialStance,
-      trustRepairPending: lastSession.trustRepairPending,
-      turnsSinceLongTermCheck: lastSession.turnsSinceLongTermCheck,
-      shortTermDeltasSummary: lastSession.shortTermDeltasSummary,
-    );
-    // Apply legacy migration (if needed) after load. (Card-seed bypass note: this *10 + migrate
-    // path is exclusively for legacy persisted sessions from pre-±300 era; fresh card seeds use
-    // the plain seedFromCardV2OrExt at the two 1:1 ext sites above. Expanded per "related load/reset
-    // sites" requirement + keep-sync full list + setActiveCharacter:1572 + both startNew.)
-    _relationshipService.applyLegacyShortTermMigrationIfNeeded();
-    _realismEnabled = lastSession.realismEnabled;
-    _moodDecayCounter = lastSession.moodDecayCounter;
-    _characterEmotion = lastSession.characterEmotion;
-    _emotionIntensity = lastSession.emotionIntensity;
-    // Time load via extracted service (resolve + scalars; keeps load blocks in sync).
-    _timeService.loadTimeScalars(
-      timeOfDay: lastSession.timeOfDay,
-      dayCount: lastSession.dayCount,
-      startDayOfWeek: lastSession.startDayOfWeek,
-      passageOfTimeEnabled:
-          lastSession.passageOfTimeEnabled &&
-          _storageService.realismSettings.passageOfTimeDefault,
-    );
-    _nsfwService.loadNsfwScalars(
-      nsfwCooldownEnabled: lastSession.nsfwCooldownEnabled,
-      arousalLevel: lastSession.arousalLevel,
-      cooldownTurnsRemaining: lastSession.cooldownTurnsRemaining,
-    );
-    _needsSimEnabled = lastSession.needsSimEnabled;
-    if (_needsSimEnabled) {
-      _needsSimulation.initializeFresh();
-      final nv = lastSession.needsVector;
-      _needsSimulation.restoreFromSnapshot({
-        'vector': (nv is String && nv.isNotEmpty)
-            ? (jsonDecode(nv) as Map).cast<String, int>()
-            : <String, int>{},
-      });
-    } else {
-      _needsSimulation.clearVector();
-    }
-
-    // Re-sync from the character's current setting so that toggling
-    // "Enjoys low hygiene" on the character affects existing chats on next load.
-    _enjoysLowHygiene =
-        _activeCharacter?.frontPorchExtensions?.enjoysLowHygiene ?? false;
-
-    _needsSimulation.resetBuffers();
-    // trust/fixation/spatial/pending/affection/tiers already loaded via _relationshipService.loadScalars above.
-    debugPrint(
-      '[ChatService] _loadLastSession: Loaded session with arousal=${_nsfwService.arousalLevel}, fixation=${_relationshipService.activeFixation}/${_relationshipService.fixationLifespan}',
-    );
-    _chaosModeService.loadScalars(
-      modeEnabled: lastSession.chaosModeEnabled,
-      pressure: lastSession.chaosPressure,
-    );
-
-    // Realism Engine 2.0 Compatibility Migration (delegated to service). (Card-seed bypass:
-    // this legacy path for old persisted data; fresh 1:1 card ext seeds use seedFromCardV2OrExt
-    // (plain) at setActive/startNew 1:1 sites. See expanded keep-sync + setActiveCharacter:1572.)
-    _relationshipService.applyLegacyShortTermMigrationIfNeeded();
-    if (_relationshipService.affectionScore != lastSession.affectionScore ||
-        _relationshipService.relationshipTier != lastSession.relationshipTier) {
-      debugPrint(
-        '[Realism] Legacy session migrated to REv2 scales (loadLast).',
-      );
-    }
-
-    // v30: Load live per-character group realism/needs (bond/trust/emotion/fixation/arousal/relationships/needs)
-    // from the session column (or fall back to group defaults). Must happen for group entry paths
-    // so that _groupRealism is populated before any eval, prompt injection, or UI read.
-    if (_activeGroup != null) {
-      _loadGroupRealismStateFromSession(lastSession);
-    } else {
-      // 1:1 session: the group realism column ('{}' for plain sessions) may
-      // carry persisted Scene Guest (Lite NPC) dbIds. Tolerant of legacy/empty.
-      _loadSceneGuestsFromSession(lastSession);
-    }
-
-    // Load per-session evolution (1:1 mode only — group is handled by _loadGroupEvolvedFields)
-    if (_activeCharacter != null) {
-      final charId = _getCharacterIdFromCard(_activeCharacter!);
-      _evolvedPersonalities[charId] = lastSession.evolvedPersonality;
-      _evolvedScenarios[charId] = lastSession.evolvedScenario;
-      _characterEvolutionCount = lastSession.evolutionCount;
-      _groupEvolutionCounts[charId] = lastSession.evolutionCount;
-    }
-
-    // Load messages
-    // Zero secondary objective flags in loaded path of _loadLast (before callers do _loadActiveObjectives / _loadObjectivesForCurrentSpeaker); incomplete zeroing hygiene.
-    _activeObjectives = [];
-    _messagesSinceLastCheck = 0;
-    _isCheckingCompletion = false;
-    _summaryPaused =
-        false; // explicit secondary zero for _summaryPaused (symmetric; _loadLast empty/loaded hygiene)
-    _isSummaryGenerating =
-        false; // secondary flag zero for summary_service (stateless/prompt-only; incomplete zeroing ... now complete)
-    _userMessagesSinceLastPeriodicEval = 0;
-    _isExtractingFacts =
-        false; // secondary fact flag + counter zero in _loadLast loaded path (incomplete zeroing ... now complete; fact_extraction)
-    _isEvolvingCharacter = false;
-    _evolutionStatus = '';
-    _evolutionError =
-        ''; // explicit evo flag/status/error zero in _loadLast loaded path (incomplete zeroing ... now complete; evolution_service (stateless or prompt-only; no reset calls needed))
-    try {
-      final dbMessages = await _db.getMessagesForSession(_currentSessionId!);
-      debugPrint(
-        '[ChatService] 🟢 _loadLastSession: loading ${dbMessages.length} '
-        'messages for session $_currentSessionId',
-      );
-      _messages.clear();
-      for (final m in dbMessages) {
-        List<String> swipes;
-        try {
-          swipes = List<String>.from(jsonDecode(m.swipes));
-        } catch (_) {
-          swipes = [''];
-        }
-        List<int> swipeDurations;
-        try {
-          swipeDurations = List<int>.from(
-            (jsonDecode(m.swipeDurations) as List).map(
-              (e) => (e as num).toInt(),
-            ),
-          );
-        } catch (_) {
-          swipeDurations = [0];
-        }
-
-        final safeSwipeIndex =
-            (m.swipeIndex >= 0 && m.swipeIndex < swipes.length)
-            ? m.swipeIndex
-            : 0;
-
-        _messages.add(
-          ChatMessage(
-            text: swipes.isNotEmpty ? swipes[safeSwipeIndex] : '',
-            sender: m.sender,
-            isUser: m.isUser,
-            characterId: m.characterId,
-            swipes: swipes,
-            swipeIndex: safeSwipeIndex,
-            swipeDurations: swipeDurations,
-            metadata: m.metadata != null
-                ? Map<String, dynamic>.from(jsonDecode(m.metadata!))
-                : null,
-            swipeMetadata: m.swipeMetadata != null
-                ? (jsonDecode(m.swipeMetadata!) as List<dynamic>)
-                      .map(
-                        (e) => e != null
-                            ? Map<String, dynamic>.from(e as Map)
-                            : null,
-                      )
-                      .toList()
-                : null,
-          ),
-        );
-      }
-
-      if (_messages.isNotEmpty) {
-        _lorebookScanner.scanLorebook(_messages.last.text);
-      }
-    } catch (e) {
-      print('Error loading chat session: $e');
-    }
-  }
-
-  /// Get sessions for a given character/group ID without setting it as active.
-  Future<List<Map<String, dynamic>>> getSessionsForId(String charId) async {
-    // Determine if this is a group or character
-    List<Session> dbSessions;
-    if (charId.startsWith('group_')) {
-      final groupId = charId.replaceFirst('group_', '');
-      dbSessions = await _db.getSessionsForGroup(groupId);
-    } else {
-      // Find character by imagePath basename
-      final allChars = await _db.getAllCharacters();
-      final match = allChars.where((c) {
-        if (c.imagePath == null) return false;
-        return path.basenameWithoutExtension(c.imagePath!) == charId;
-      }).firstOrNull;
-      if (match != null) {
-        dbSessions = await _db.getSessionsForCharacter(match.id);
-      } else {
-        dbSessions = [];
-      }
-    }
-
-    List<Map<String, dynamic>> sessions = [];
-    for (final s in dbSessions) {
-      final timestamp = int.tryParse(s.id) ?? 0;
-      final date = DateTime.fromMillisecondsSinceEpoch(timestamp);
-
-      // Get message count and preview
-      final msgs = await _db.getMessagesForSession(s.id);
-      String preview = 'New Conversation';
-      if (s.name != null && s.name!.isNotEmpty) {
-        preview = s.name!;
-      } else if (msgs.length > 1) {
-        // Use second message text as preview
-        try {
-          final swipes = List<String>.from(jsonDecode(msgs[1].swipes));
-          preview = swipes.isNotEmpty ? swipes[msgs[1].swipeIndex] : '';
-          if (preview.length > 50) preview = '${preview.substring(0, 50)}...';
-        } catch (_) {}
-      }
-
-      sessions.add({
-        'id': s.id,
-        'date': date,
-        'preview': preview,
-        'message_count': msgs.length,
-        'user_message_count': msgs.where((m) => m.isUser).length,
-        if (s.name != null) 'session_name': s.name,
-        if (s.description != null) 'session_description': s.description,
-        if (s.parentSession != null) 'parent_session': s.parentSession,
-        if (s.forkIndex != null) 'fork_index': s.forkIndex,
-      });
-    }
-
-    sessions.sort(
-      (a, b) => (b['date'] as DateTime).compareTo(a['date'] as DateTime),
-    );
-    return sessions;
-  }
-
-  Future<List<Map<String, dynamic>>> getSessions() async {
-    if (_activeCharacter == null && _activeGroup == null) return [];
-    final charId = _getCharacterId();
-    return getSessionsForId(charId);
-  }
-
-  Future<void> loadSession(String sessionId) async {
-    if (_activeCharacter == null && _activeGroup == null) return;
-
-    final session = await _db.getSessionById(sessionId);
-    if (session == null) return;
-
-    if (session.userPersonaId != null) {
-      await _userPersonaService.setActivePersona(session.userPersonaId!);
-    }
-
-    try {
-      final dbMessages = await _db.getMessagesForSession(sessionId);
-      debugPrint(
-        '[ChatService] 🟢 loadSession: loading ${dbMessages.length} '
-        'messages for session $sessionId',
-      );
-      _messages.clear();
-      for (final m in dbMessages) {
-        List<String> swipes;
-        try {
-          swipes = List<String>.from(jsonDecode(m.swipes));
-        } catch (_) {
-          swipes = [''];
-        }
-        List<int> swipeDurations;
-        try {
-          swipeDurations = List<int>.from(
-            (jsonDecode(m.swipeDurations) as List).map(
-              (e) => (e as num).toInt(),
-            ),
-          );
-        } catch (_) {
-          swipeDurations = [0];
-        }
-
-        final safeSwipeIndex =
-            (m.swipeIndex >= 0 && m.swipeIndex < swipes.length)
-            ? m.swipeIndex
-            : 0;
-
-        _messages.add(
-          ChatMessage(
-            text: swipes.isNotEmpty ? swipes[safeSwipeIndex] : '',
-            sender: m.sender,
-            isUser: m.isUser,
-            characterId: m.characterId,
-            swipes: swipes,
-            swipeIndex: safeSwipeIndex,
-            swipeDurations: swipeDurations,
-            metadata: m.metadata != null
-                ? Map<String, dynamic>.from(jsonDecode(m.metadata!))
-                : null,
-            swipeMetadata: m.swipeMetadata != null
-                ? (jsonDecode(m.swipeMetadata!) as List<dynamic>)
-                      .map(
-                        (e) => e != null
-                            ? Map<String, dynamic>.from(e as Map)
-                            : null,
-                      )
-                      .toList()
-                : null,
-          ),
-        );
-      }
-
-      // Post-load sanitization: force valid swipe indices and clamp absurdly long fixation text.
-      // This protects against any legacy corrupted rows or previous buggy saves, even if the
-      // individual message constructors already clamp.
-      for (final msg in _messages) {
-        if (msg.swipeIndex < 0 || msg.swipeIndex >= msg.swipes.length) {
-          msg.swipeIndex = 0;
-        }
-      }
-
-      // The fixation coming out of the LLM can sometimes be a full paragraph instead of a short topic.
-      // Truncate it to keep the UI and prompts sane.
-      _relationshipService.sanitizeFixationIfTooLong();
-
-      // ── Hydrate hidden group state checkpoint (DB-free: realism + per-char notes) ──
-      // The sentinel is stored as the last message for durability but must be
-      // stripped from the in-memory list so the UI and prompt builders never see it.
-      // (v30: _hydrateGroupRealismCheckpointIfPresent removed — state now loads from DB column)
-
-      _currentSessionId = sessionId;
-      // Scene Guests are per-session. Without this, switching to a different
-      // session via the history picker leaves the PREVIOUS session's guests
-      // (and their evolution/detection state) in place — they keep chiming in
-      // and get re-persisted into the loaded session's blob, contaminating it,
-      // while this session's own guests are never restored. Mirror the
-      // _loadLastSession 1:1 branch: full reset, then load this session's blob.
-      if (_activeGroup == null) {
-        _pendingGuestDeparture = null;
-        _pendingGuestPickerFilter = null;
-        _resetGuestActivityState();
-        _userMessagesSinceLastCastScan = 0;
-        _pendingGuestDetection = null;
-        _offeredOrIgnoredGuestNames.clear();
-        // (clears + restores _sceneGuestIds/_sceneGuestCards + guest evolution)
-        _loadSceneGuestsFromSession(session);
-      }
-      _authorNote = session.authorNote;
-      _authorNoteStrength = session.authorNoteDepth;
-      _summary = session.summary ?? '';
-      _summaryLastIndex = session.summaryLastIndex ?? 0;
-      _summaryPaused =
-          false; // explicit secondary zero for _summaryPaused on loadSession loaded path (incomplete zeroing ... now complete; see keep-sync + summary_service)
-      _isSummaryGenerating =
-          false; // secondary zero for flag on loadSession loaded (symmetric)
-      _userMessagesSinceLastPeriodicEval = 0;
-      _isExtractingFacts =
-          false; // secondary fact flag + counter zero on loadSession loaded (symmetric; incomplete zeroing ... now complete; fact_extraction)
-      _isEvolvingCharacter = false;
-      _evolutionStatus = '';
-      _evolutionError =
-          ''; // explicit evo flag/status/error zero on loadSession loaded (symmetric; incomplete zeroing ... now complete; evolution_service (stateless or prompt-only; no reset calls needed))
-      _sessionName = session.name;
-      _sessionDescription = session.description;
-      _parentSessionId = session.parentSession;
-      _forkIndex = session.forkIndex;
-      // Relationship load + tier calc + legacy migration via service.
-      _relationshipService.loadScalars(
-        affectionScore: session.affectionScore,
-        longTermScore: session.longTermScore,
-        trustLevel: session.trustLevel,
-        activeFixation: session.activeFixation,
-        fixationLifespan: session.fixationLifespan,
-        spatialStance: session.spatialStance,
-        trustRepairPending: session.trustRepairPending,
-        turnsSinceLongTermCheck: session.turnsSinceLongTermCheck,
-        shortTermDeltasSummary: session.shortTermDeltasSummary,
-      );
-      _relationshipService.applyLegacyShortTermMigrationIfNeeded();
-      // (Card-seed bypass: legacy *10 migration only; see the two ext-seed sites + prior load sites
-      // for full "card seeds authored on current ±300" + keep-sync lists + relationship leaf.)
-
-      // counters already via loadScalars on service.
-      _realismEnabled = session.realismEnabled;
-      _moodDecayCounter = session.moodDecayCounter;
-      _characterEmotion = session.characterEmotion;
-      _emotionIntensity = session.emotionIntensity;
-      // Time load via extracted service (resolve + scalars; keeps group load blocks in sync).
-      _timeService.loadTimeScalars(
-        timeOfDay: session.timeOfDay,
-        dayCount: session.dayCount,
-        startDayOfWeek: session.startDayOfWeek,
-        passageOfTimeEnabled:
-            session.passageOfTimeEnabled &&
-            _storageService.realismSettings.passageOfTimeDefault,
-      );
-      _nsfwService.loadNsfwScalars(
-        nsfwCooldownEnabled: session.nsfwCooldownEnabled,
-        arousalLevel: session.arousalLevel,
-        cooldownTurnsRemaining: session.cooldownTurnsRemaining,
-      );
-      _needsSimEnabled = session.needsSimEnabled;
-      if (_needsSimEnabled) {
-        _needsSimulation.initializeFresh();
-        final nv = session.needsVector;
-        _needsSimulation.restoreFromSnapshot({
-          'vector': (nv is String && nv.isNotEmpty)
-              ? (jsonDecode(nv) as Map).cast<String, int>()
-              : <String, int>{},
-        });
-      } else {
-        _needsSimulation.clearVector();
-      }
-
-      // Re-sync from the character's current setting so toggling
-      // "Enjoys low hygiene" affects existing chats on load.
-      _enjoysLowHygiene =
-          _activeCharacter?.frontPorchExtensions?.enjoysLowHygiene ?? false;
-
-      _needsSimulation.resetBuffers();
-      // trust/fixation etc already via _relationshipService.loadScalars above.
-
-      // Load per-session evolution (1:1 mode only — group handled by _loadGroupEvolvedFields)
-      if (_activeCharacter != null) {
-        final charId = _getCharacterIdFromCard(_activeCharacter!);
-        _evolvedPersonalities[charId] = session.evolvedPersonality;
-        _evolvedScenarios[charId] = session.evolvedScenario;
-        _characterEvolutionCount = session.evolutionCount;
-        _groupEvolutionCounts[charId] = session.evolutionCount;
-      }
-
-      // Per-session generation parameter overrides (v22) — loaded via raw SQL
-      // so this works even before build_runner regenerates database.g.dart.
-      try {
-        final genRows = await _db
-            .customSelect(
-              'SELECT generation_settings FROM sessions WHERE id = ?',
-              variables: [drift.Variable(sessionId)],
-            )
-            .get();
-        final genJson = genRows.isNotEmpty
-            ? genRows.first.read<String?>('generation_settings')
-            : null;
-        _sessionGenSettings = ChatGenerationSettings.fromJsonString(genJson);
-      } catch (_) {
-        _sessionGenSettings = ChatGenerationSettings();
-      }
-
-      if (_messages.isNotEmpty) {
-        _lorebookScanner.scanLorebook(_messages.last.text);
-      }
-      notifyListeners();
-    } catch (e) {
-      print('Error loading session $sessionId: $e');
-    }
-  }
-
-  /// Rename a session.
-  Future<void> renameSession(String sessionId, String name) async {
-    final session = await _db.getSessionById(sessionId);
-    if (session == null) return;
-
-    await _db.updateSession(
-      SessionsCompanion(
-        id: drift.Value(sessionId),
-        name: drift.Value(name.isEmpty ? null : name),
-        updatedAt: drift.Value(DateTime.now()),
-      ),
-    );
-
-    // Update in-memory if this is the current session
-    if (sessionId == _currentSessionId) {
-      _sessionName = name.isEmpty ? null : name;
-      notifyListeners();
-    }
-  }
-
-  /// Update the description of a session.
-  Future<void> updateSessionDescription(
-    String sessionId,
-    String description,
-  ) async {
-    final session = await _db.getSessionById(sessionId);
-    if (session == null) return;
-
-    await _db.updateSession(
-      SessionsCompanion(
-        id: drift.Value(sessionId),
-        description: drift.Value(description.isEmpty ? null : description),
-        updatedAt: drift.Value(DateTime.now()),
-      ),
-    );
-
-    // Update in-memory if this is the current session
-    if (sessionId == _currentSessionId) {
-      _sessionDescription = description.isEmpty ? null : description;
-      notifyListeners();
-    }
-  }
-
-  /// Create a new session by forking from message at [messageIndex].
-  /// Copies messages 0..messageIndex into a new session and switches to it.
-  Future<void> forkFromMessage(int messageIndex) async {
-    if ((_activeCharacter == null && _activeGroup == null) ||
-        _currentSessionId == null) {
-      return;
-    }
-    if (messageIndex < 0 || messageIndex >= _messages.length) return;
-
-    final oldSessionId = _currentSessionId!;
-    final forkedMessages = _messages
-        .sublist(0, messageIndex + 1)
-        .map(
-          (m) => ChatMessage(
-            text: m.text,
-            sender: m.sender,
-            isUser: m.isUser,
-            characterId: m.characterId,
-            swipes: List.from(m.swipes),
-            swipeIndex: (m.swipeIndex >= 0 && m.swipeIndex < m.swipes.length)
-                ? m.swipeIndex
-                : 0,
-            swipeDurations: List.from(m.swipeDurations),
-            metadata: m.metadata != null
-                ? Map<String, dynamic>.from(m.metadata!)
-                : null,
-            swipeMetadata: m.swipeMetadata
-                .map((e) => e != null ? Map<String, dynamic>.from(e) : null)
-                .toList(),
-          ),
-        )
-        .toList();
-
-    debugPrint(
-      '[ChatService] 🟡 forkSession: clearing messages for fork at index $messageIndex',
-    );
-    _messages.clear();
-    _messages.addAll(forkedMessages);
-    _currentSessionId = DateTime.now().millisecondsSinceEpoch.toString();
-    _parentSessionId = oldSessionId;
-    _forkIndex = messageIndex;
-    _sessionGenSettings = _sessionGenSettings
-        .copy(); // inherit parent's overrides
-    _summary = '';
-    _summaryLastIndex = 0;
-    _summaryPaused =
-        false; // explicit secondary zero for _summaryPaused (symmetric to generating; fork hygiene + incomplete zeroing now complete)
-    _isSummaryGenerating =
-        false; // zero secondary flag on fork (new branch hygiene, matches summary scalar reset)
-    _userMessagesSinceLastPeriodicEval = 0;
-    _isExtractingFacts =
-        false; // secondary fact flag + counter zero on fork (new branch hygiene + incomplete zeroing now complete; fact_extraction)
-    _isEvolvingCharacter = false;
-    _evolutionStatus = '';
-    _evolutionError =
-        ''; // explicit evo flag/status/error zero on fork (new branch hygiene + incomplete zeroing now complete; evolution_service (stateless or prompt-only; no reset calls needed))
-
-    // Time-Travel Restoration
-    if (_messages.isNotEmpty) {
-      _restoreRealismStateFromMessage(_messages.last);
-    }
-
-    await _saveChat();
-    notifyListeners();
-  }
-
-  Future<void> startNewChat() async {
-    if (_activeCharacter == null && _activeGroup == null) return;
-
-    debugPrint(
-      '[startNewChat] START: arousal=${_nsfwService.arousalLevel}, fixation=${_relationshipService.activeFixation}/${_relationshipService.fixationLifespan}',
-    );
-
-    // Refresh _activeCharacter from the repository so we pick up any edits
-    // made in the character editor (personality, description, etc.)
-    if (_activeCharacter != null && _characterRepository != null) {
-      final freshChar = _characterRepository!.characters
-          .cast<CharacterCard?>()
-          .firstWhere(
-            (c) => c!.dbId == _activeCharacter!.dbId,
-            orElse: () => null,
-          );
-      if (freshChar != null) {
-        // Preserve any runtime-loaded extensions if the repository instance lacks them
-        final existingExt = _activeCharacter!.frontPorchExtensions;
-        final existingRaw = _activeCharacter!.rawExtensions;
-
-        _activeCharacter = freshChar;
-
-        if (existingExt != null &&
-            _activeCharacter!.frontPorchExtensions == null) {
-          _activeCharacter!.frontPorchExtensions = existingExt;
-          _activeCharacter!.rawExtensions = existingRaw;
-          debugPrint(
-            '[startNewChat] Preserved existing extensions during character refresh',
-          );
-        } else if (existingExt == null &&
-            _activeCharacter!.frontPorchExtensions == null) {
-          debugPrint(
-            '[startNewChat] DEBUG: existingExt was null AND freshChar had null extensions.',
-          );
-        }
-      } else {
-        debugPrint(
-          '[startNewChat] DEBUG: freshChar was null (repository lookup failed).',
-        );
-      }
-    }
-
-    // Fallback: If extensions are STILL missing, forcefully reload from PNG.
-    // This catches edge cases where repository/memory loses sync with the file.
-    if (_activeCharacter != null &&
-        _activeCharacter!.frontPorchExtensions == null) {
-      debugPrint(
-        '[startNewChat] DEBUG: Extensions are missing. Attempting PNG fallback.',
-      );
-      if (_activeCharacter!.imagePath != null) {
-        debugPrint(
-          '[startNewChat] DEBUG: Image path exists: ${_activeCharacter!.imagePath}',
-        );
-        try {
-          final v2Service = V2CardService();
-          final reloaded = await v2Service.readCard(
-            _activeCharacter!.imagePath!,
-          );
-          if (reloaded == null) {
-            debugPrint(
-              '[startNewChat] DEBUG: readCard returned null. Failed to parse PNG.',
-            );
-          } else if (reloaded.frontPorchExtensions != null) {
-            _activeCharacter!.frontPorchExtensions =
-                reloaded.frontPorchExtensions;
-            _activeCharacter!.rawExtensions = reloaded.rawExtensions;
-            debugPrint(
-              '[startNewChat] Force-reloaded frontPorchExtensions from PNG',
-            );
-          } else {
-            debugPrint(
-              '[startNewChat] DEBUG: readCard succeeded, BUT reloaded.frontPorchExtensions was NULL! Meaning the PNG file does NOT contain the front_porch extension data.',
-            );
-          }
-        } catch (e) {
-          debugPrint('[startNewChat] Force-reload failed with exception: $e');
-        }
-      } else {
-        debugPrint('[startNewChat] DEBUG: Cannot fallback. imagePath is null.');
-      }
-    }
-
-    debugPrint(
-      '[ChatService] 🟡 startNewChat: clearing messages (had ${_messages.length})',
-    );
-    _messages.clear();
-    _greetingIndex = 0;
-    // A fresh chat starts with no Scene Guests (they don't carry across sessions).
-    _clearSceneGuestEvolution(); // Phase 3: keep guest-evolution reset in sync.
-    _sceneGuestIds.clear();
-    _sceneGuestCards.clear();
-    _pendingGuestDeparture = null;
-    _pendingGuestPickerFilter = null;
-    _resetGuestActivityState();
-    // Phase 2 cast detection: reset the scan cadence + pending/debounce state
-    // for the new 1:1 context (kept in sync with the Scene Guest clears).
-    _userMessagesSinceLastCastScan = 0;
-    _pendingGuestDetection = null;
-    _offeredOrIgnoredGuestNames.clear();
-    _summary = '';
-    _summaryLastIndex = 0;
-    _summaryPaused =
-        false; // explicit secondary zero for _summaryPaused (symmetric; startNew 1:1/ext-seed branch + incomplete zeroing ... now complete)
-    _isSummaryGenerating =
-        false; // explicit in startNewChat 1:1/ext-seed branch (both startNew explicit + incomplete zeroing... now complete (see CLAUDE.md); summary_service) + "needsSimulation. (reason support kept for Director chips) ; cleared via sim initializeFresh/clearVector/resetBuffers on all paths; now complete in both branches)"
-
-    // Explicitly clear any prior branching/fork metadata. A "New Chat" is
-    // never a branch/fork from a previous session. This prevents stale
-    // _parentSessionId / _forkIndex (from a previous branched chat or
-    // different character) from being written into the brand-new session
-    // record via _saveChat(), which was causing new chats to incorrectly
-    // show "Branched at message #NNNN" in history lists even for characters
-    // with no prior chats.
-    _parentSessionId = null;
-    _forkIndex = null;
-
-    // Mark this as a new chat to prevent memory retrieval
-    _isNewChat = true;
-    debugPrint('[startNewChat] Marked as new chat - memories will be filtered');
-
-    // Save the current session (preserves objectives for this session)
-    if (_currentSessionId != null) {
-      await _saveChat();
-    }
-
-    // Clear objectives for fresh session start
-    _activeObjectives = [];
-    _messagesSinceLastCheck = 0;
-    _isCheckingCompletion =
-        false; // see decl + keep reset blocks (incomplete zeroing... now complete (see CLAUDE.md); explicit in both startNew branches)
-    _userMessagesSinceLastPeriodicEval = 0;
-    _isExtractingFacts =
-        false; // explicit secondary fact flag + counter zero in startNew 1:1/ext-seed branch (both startNew explicit + incomplete zeroing ... now complete; fact_extraction)
-    _isEvolvingCharacter = false;
-    _evolutionStatus = '';
-    _evolutionError =
-        ''; // explicit evo flag/status/error zero in startNew 1:1/ext-seed branch (both startNew explicit + incomplete zeroing ... now complete; evolution_service (stateless or prompt-only; no reset calls needed))
-
-    // Create new session ID for the new chat
-    _currentSessionId = DateTime.now().millisecondsSinceEpoch.toString();
-
-    // Clear memory sources to prevent old memories from being retrieved
-    // Cross-character memory can still be re-selected by user after new chat starts
-    if (_activeCharacter?.dbId != null) {
-      try {
-        await _db.updateCharacter(
-          CharactersCompanion(
-            id: drift.Value(_activeCharacter!.dbId!),
-            memorySources: drift.Value('[]'),
-          ),
-        );
-        debugPrint('[startNewChat] Cleared memory sources from DB');
-      } catch (e) {
-        debugPrint('[startNewChat] Failed to clear memory sources: $e');
-      }
-    }
-
-    // Seed Realism Engine state from V2.5 card extensions for 1:1 mode only,
-    // ensuring realism settings persist across chat sessions (group mode handled elsewhere)
-    if (_activeCharacter != null && _activeGroup == null) {
-      final extSeed =
-          _activeCharacter!.frontPorchExtensions ?? FrontPorchExtensions();
-
-      _realismEnabled = extSeed.realismEnabled;
-      // Card-seed bypass (rec 1 from PR #47; keeps startNewChat parity with setActive ext seed):
-      // use seedFromCardV2OrExt (plain .clamp only, no _migrate*) because V2.5 cards + creator
-      // author on current ±300 scale. (The old "Migration + seed" comment + call was the source
-      // of the doubling regression on fresh 1:1 New Chat.) Migration stays exclusively on legacy
-      // persisted session paths (_loadLastSession + loadScalars(migrate*) + applyLegacy... at 3 sites).
-      // 1:1 vs group parity: group seeding paths untouched (resetForFresh + per-speaker load/save scalars).
-      // See relationship_service.dart (seedFromCardV2OrExt + public migrate docs) + expanded
-      // "keep reset blocks in sync" + "incomplete zeroing... now complete (see CLAUDE.md)"
-      // (see CLAUDE.md full list + incomplete zeroing hygiene; buffer removal complete)
-      _relationshipService.seedFromCardV2OrExt(
-        shortTermBond: extSeed.shortTermBond,
-        longTermBond: extSeed.longTermBond,
-        trustLevel: extSeed.trustLevel,
-      );
-      _expressionService.resetForFreshChat();
-      // Lorebook trigger reset via extracted service (keeps the keep-sync reset sites correct
-      // without god privates; now includes startNewChat 1:1 ext-seed path to prevent bleed of prior
-      // isTriggered/remainingDepth into fresh New Chat for 1:1; constants skipped. See setActiveCharacter:1572
-      // + "incomplete zeroing of secondary realism configuration fields" briefing pattern (cross-ref step6 nsfw).
-      // See lorebook_scanner.dart and "keep reset blocks" comments (now lists needs/chaos/... + leaves (see CLAUDE.md for full; incomplete zeroing now complete) + needs_impact_evaluator (stateless or prompt-only; no reset calls needed)). (cross-ref setActiveCharacter:1572 etc; card-seed bypass hygiene added here too)
-      _lorebookScanner.resetLorebookTriggerState();
-      // Time seed via extracted service (keeps startNewChat / setActive / ext-seed blocks in sync).
-      _timeService.seedFromV2OrExt(
-        dayCount: extSeed.dayCount.clamp(1, 9999),
-        timeOfDay: extSeed.timeOfDay,
-        passageOfTimeEnabled:
-            extSeed.passageOfTimeEnabled &&
-            _storageService.realismSettings.passageOfTimeDefault,
-      );
-      _characterEmotion = extSeed.characterEmotion;
-      _emotionIntensity = extSeed.emotionIntensity;
-      _nsfwService.seedFromV2OrExt(
-        nsfwCooldownEnabled: extSeed.nsfwCooldownEnabled,
-      );
-      _chaosModeService.seedFromGroupOrExt(extSeed.chaosModeEnabled, false);
-      _needsSimEnabled = extSeed.needsSimEnabled;
-      _enjoysLowHygiene = extSeed.enjoysLowHygiene;
-      if (_needsSimEnabled) {
-        // Fresh chat / new session: seed from card baselines (falls back to
-        // needDefaults when the card has no baselines).
-        _needsSimulation.initializeFreshWithDefaults({
-          'hunger': extSeed.needsBaselineHunger,
-          'bladder': extSeed.needsBaselineBladder,
-          'energy': extSeed.needsBaselineEnergy,
-          'social': extSeed.needsBaselineSocial,
-          'fun': extSeed.needsBaselineFun,
-          'hygiene': extSeed.needsBaselineHygiene,
-          'comfort': extSeed.needsBaselineComfort,
-        });
-      } else {
-        _needsSimulation.clearVector();
-      }
-      _needsSimulation.resetBuffers();
-      // needs_impact_evaluator is stateless/prompt-only (no reset calls needed on it;
-      // see full list in "keep reset blocks in sync" comments + cross-ref setActiveCharacter:1572 + fact_extraction (stateless or prompt-only; no reset calls needed) + evolution_service (stateless or prompt-only; no reset calls needed) + realism_verification (stateless or prompt-only; no reset calls needed) + " ; read live from ext on active/group speaker; incomplete zeroing... now complete (see CLAUDE.md) (no extra scalar)") + "needsSimulation. (reason support kept for Director chips) ; cleared via sim initializeFresh/clearVector/resetBuffers on all paths; now complete)".
-
-      // pending covered by relationship service reset in the ext-seed or non-ext paths below.
-      // Always reset per-chat runtime realism fields (arousal/fixation/cooldowns) for a fresh
-      // session started via explicit New Chat. ... (See also: matching full reset in setActiveCharacter
-      // and the cross-sync comment there; load*Session, deleteSession→startNewChat, setActiveGroup defensive zero.
-      // Needs vector/buffers reset via _needsSimulation also kept in sync across sites.)
-      // Time secondary fields (passage/anchor/turns/day/time) zeroed via _timeService.resetForFreshChat in non-ext path + _load empty + setActiveGroup.
-      // Nsfw runtime (arousal/cooldown) zeroed via service for fresh (see resetRuntime + resetForFreshChat).
-      // Declarative initial bond/trust/emotion/day etc are already seeded above from the card's
-      // FrontPorchExtensions (or defaults). The old hasFrontPorchExtensions preserve here
-      // was the source of fixation bleed on "New Chat" for cards that had any FP ext object.
-      // Expression (manual/caches/onnx/lastAvatar) reset via service for no-bleed (new for step 4).
-      // Lorebook (non-const isTriggered/remainingDepth) reset via scanner in both ext-seed (1:1) and non-ext (group/0-session) paths of startNewChat (added for hygiene to match briefing "every keep-sync" + setActiveCharacter etc; prevents bleed into greetings/scans).
-      debugPrint(
-        '[startNewChat] Resetting runtime arousal/fixation + transients for fresh chat (was: arousal=${_nsfwService.arousalLevel}, fixation=${_relationshipService.activeFixation}/${_relationshipService.fixationLifespan})',
-      );
-      _nsfwService.resetRuntimeArousalAndCooldown();
-      debugPrint(
-        '[startNewChat] After reset: arousal=${_nsfwService.arousalLevel}, fixation=${_relationshipService.activeFixation}/${_relationshipService.fixationLifespan}',
-      );
-
-      // Recalculate tiers from seeded scores (only needed for realism-enabled chars)
-      if (_realismEnabled) {
-        // Tiers are maintained inside service after seed; no direct _calculate here.
-      }
-
-      // Seed initial quest/task as a primary objective
-      if (extSeed.currentTask.isNotEmpty) {
-        // Defer so the session ID is ready before the DB write
-        Future.microtask(() async {
-          await setObjective(extSeed.currentTask, isPrimary: true);
-          debugPrint(
-            '[ChatService] V2.5 seeded initial task: ${extSeed.currentTask}',
-          );
-        });
-      }
-    } else {
-      // Group mode or no active character: reset to defaults but preserve existing extensions-based values
-      // (pending covered by service.resetForFreshChat below)
-
-      if (_activeGroup == null && _messages.isNotEmpty) {
-        // Will be populated later with greeting in non-group modes
-        // Preserve realism state for proper post-greeting eval (don't reset here)
-      } else {
-        // Relationship + Expression + Time + Nsfw reset via service helpers (keeps reset blocks in sync with setActiveCharacter:1572 etc / _loadLast empty / setActiveGroup / startNew ext-seed; see "incomplete zeroing... now complete (see CLAUDE.md)" + full list in "keep reset blocks" comments including + llm_eval_engine (stateless or prompt-only; no reset calls needed; incomplete zeroing... now complete (see CLAUDE.md)) + needs_impact_evaluator (stateless or prompt-only; no reset calls needed) + realism_evals (stateless or prompt-only; no reset calls needed)). (cross-ref setActiveCharacter:1572 etc)
-        // Time now explicitly reset in group 0-session/empty paths + setActiveGroup defensive + _loadLast empty (cross-check needs bugfix hygiene).
-        _relationshipService.resetForFreshChat();
-        _expressionService.resetForFreshChat();
-        _timeService.resetForFreshChat();
-        _nsfwService.resetForFreshChat();
-        // Lorebook trigger reset via extracted service (keeps reset blocks in sync with setActiveCharacter:1572 / _loadLast empty / setActiveGroup / startNew ext-seed; see "incomplete zeroing of secondary ... on 0-session/new-character/group" + startNew 1:1+group now complete + full list in keep-sync comments incl llm_eval_engine). (cross-ref setActiveCharacter:1572 etc)
-        // See "keep reset blocks in sync" comments (setActiveGroup, startNewChat, load* , setActive* all must hit this; now includes needs/chaos/... + leaves (see CLAUDE.md for full; incomplete zeroing now complete) + " )" for group/0-session/new-chat hygiene; incomplete zeroing now complete).
-        // (cross-ref setActiveCharacter:1572)
-        _lorebookScanner.resetLorebookTriggerState();
-        // Don't touch dayCount/time etc directly — seeded from extensions or loaded session (or reset above for fresh no-ext path).
-        // Time reset helper kept in sync with other blocks.
-        // needs_impact_evaluator (stateless/prompt-only; no reset calls needed) covered in keep-sync lists.
-
-        // Explicit zero for secondary config flags in group/non-ext/0-session/new-chat path (keeps "incomplete zeroing... now complete (see CLAUDE.md)" true in *code* not just comments; matches ext-seed 1:1 + setActiveCharacter + setActiveGroup defensive; cross-ref setActiveCharacter:1572 + full list in keep-sync comments incl + needs_impact_evaluator (stateless or prompt-only; no reset calls needed) + realism_evals (stateless or prompt-only; no reset calls needed) + " ; authority mode for needs; hygiene listed, no code zero line)") + "needsSimulation. (reason support kept for Director chips) ; cleared via sim initializeFresh/clearVector/resetBuffers on all paths; now complete)".
-        _needsSimEnabled = false;
-        _enjoysLowHygiene = false;
-        _needsSimulation.clearVector();
-        _needsSimulation.resetBuffers();
-        _activeObjectives = [];
-        _messagesSinceLastCheck = 0;
-        _isCheckingCompletion =
-            false; // explicit in non-ext/group/0-session else branch of startNew (both branches now; incomplete zeroing ... now complete)
-        _summaryPaused =
-            false; // explicit secondary zero for _summaryPaused (symmetric to generating; non-ext/group/0-session startNew path + now complete)
-        _isSummaryGenerating =
-            false; // explicit secondary zero in startNew non-ext/group/0-session path (both branches + now complete for summary flag too)
-        _userMessagesSinceLastPeriodicEval = 0;
-        _isExtractingFacts =
-            false; // explicit secondary fact flag + counter zero in startNew non-ext/group/0-session path (both branches + now complete for fact flag/counter; fact_extraction)
-        _isEvolvingCharacter = false;
-        _evolutionStatus = '';
-        _evolutionError =
-            ''; // explicit evo flag/status/error zero in startNew non-ext/group/0-session path (both branches + now complete for evo flag; evolution_service (stateless or prompt-only; no reset calls needed) + " )") + "needsSimulation. (reason support kept for Director chips) ; cleared via sim initializeFresh/clearVector/resetBuffers on all paths; now complete)"
-      }
-    }
-
-    // Explicit flag + cadence counter zero for evolution (in addition to per-branch) to keep "incomplete zeroing... now complete (see CLAUDE.md)" + both startNew explicit; evolution_service (stateless or prompt-only; no reset calls needed) + " ; no god scalar zero needed -- live ext read; see also setActiveCharacter + group 0-session paths)" + "needsSimulation. (reason support kept for Director chips) ; cleared via sim initializeFresh/clearVector/resetBuffers on all paths; now complete)".
-    // Also zero the facts counter here for symmetric hygiene on the two periodic cadence counters.
-    _userMessagesSinceLastPeriodicEval = 0;
-    _isEvolvingCharacter = false;
-    _evolutionStatus = '';
-    _evolutionError = '';
-
-    // Clear the in-memory evolution cache so the new session starts with
-    // the original (unevolved) personality/scenario. The previous session's
-    // evolved data was still live in this map. (Flags zeroed explicitly in branches + here for hygiene; see preceding comment.)
-    _evolvedPersonalities.clear();
-    _evolvedScenarios.clear();
-    _groupEvolutionCounts.clear();
-    _characterEvolutionCount = 0;
-
-    if (_activeGroup != null && _groupCharacters.isNotEmpty) {
-      // Group mode: respect explicit group.firstMessage (custom group greeting set
-      // by creator or Group Card) when present. Only fall back to the first
-      // participating character's firstMessage when the group has no custom opening.
-      String greetingText;
-      String greetingSender;
-      String? greetingCharId;
-
-      if (_activeGroup!.firstMessage.isNotEmpty) {
-        greetingText = _macroResolver.resolve(
-          _activeGroup!.firstMessage,
-          MacroContext(userName: _userPersonaService.persona.name),
-          section: 'greeting',
-        );
-        greetingSender = _activeGroup!.name;
-        greetingCharId = null;
-      } else {
-        final first = _groupCharacters.first;
-        greetingText = first.firstMessage.isNotEmpty
-            ? _buildFirstMessage(first)
-            : '';
-        greetingSender = first.name;
-        greetingCharId = _getCharacterIdFromCard(first);
-      }
-
-      if (greetingText.isNotEmpty) {
-        _messages.add(
-          ChatMessage(
-            text: greetingText,
-            sender: greetingSender,
-            isUser: false,
-            characterId: greetingCharId,
-          ),
-        );
-        _lorebookScanner.scanLorebook(_messages.last.text);
-      }
-      _groupManager?.resetTurnState();
-    } else if (_activeCharacter != null) {
-      // 1:1 mode
-      if (_activeCharacter!.firstMessage.isNotEmpty) {
-        _messages.add(
-          ChatMessage(
-            text: _buildFirstMessage(_activeCharacter!),
-            sender: _activeCharacter!.name,
-            isUser: false,
-          ),
-        );
-        _lorebookScanner.scanLorebook(_messages.last.text);
-      }
-    }
-
-    _currentSessionId = DateTime.now().millisecondsSinceEpoch.toString();
-    debugPrint(
-      '[startNewChat] BEFORE SAVE: arousal=${_nsfwService.arousalLevel}, fixation=${_relationshipService.activeFixation}/${_relationshipService.fixationLifespan}',
-    );
-    await _saveChat();
-    debugPrint(
-      '[startNewChat] AFTER SAVE: arousal=${_nsfwService.arousalLevel}, fixation=${_relationshipService.activeFixation}/${_relationshipService.fixationLifespan}',
-    );
-    notifyListeners();
-
-    // ── Post-Greeting Realism Baseline ──────────────────────────────────
-    // Always mark that a greeting was placed — even if Realism is currently off.
-    // If Realism is already on, fire immediately. Otherwise the flag will be
-    // consumed the moment the user enables Realism.
-    // Skip if character already has pre-seeded V2.5 extensions — those baseline
-    // values are intentional and should not be overwritten by auto-eval.
-    // (See also: setActiveCharacter 0-session path comment for why direct imports rely on retro path.)
-    if (_activeGroup == null &&
-        _messages.isNotEmpty &&
-        _activeCharacter!.frontPorchExtensions == null) {
-      _greetingEvalPending = true;
-      if (_realismEnabled) {
-        _runPostGreetingEval();
-      }
-    }
-  }
 
   /// Evaluates emotion + relationship baseline from the greeting message only.
   /// Runs once per new session, silently in the background.
